fix(redis db.statement): only serialize non sensitive arguments in db.statement attribute
