+++ packages/opentelemetry-redis-common/README.md
@@ -1,0 +1,7 @@
+# Redis Common Utils for OpenTelemetry redis packages
+
+This is an internal utils package used for the different redis instrumentations:
+
+1. ioredis
+2. redis
+3. redis-4
