+++ plugins/node/opentelemetry-instrumentation-redis-4/src/utils.ts
@@ -18,9 +18,9 @@ import {
   DbSystemValues,
   SemanticAttributes,
 } from '@opentelemetry/semantic-conventions';
+import { defaultDbStatementSerializer as redisSerializer } from '@opentelemetry/redis-common';
 
-export const defaultDbStatementSerializer: DbStatementSerializer = cmdName =>
-  cmdName;
+export const defaultDbStatementSerializer: DbStatementSerializer = redisSerializer;
 
 export function getClientAttributes(options: any) {
   return {
